+++ fake-cucumber/javascript/src/types.ts
@@ -1,8 +1,8 @@
 import { Readable } from 'stream'
 import { messages } from '@cucumber/messages'
 
-export type MessageNotifier = (message: messages.IEnvelope) => void
-export type AnyBody = (...args: any) => any
+export type EnvelopeListener = (envelope: messages.IEnvelope) => void
+export type AnyBody = (...args: ReadonlyArray<any>) => any
 export type Attach = (
   data: string | Buffer | Readable,
   mediaType: string