+++ react/javascript/src/CucumberQueryContext.ts
@@ -1,4 +1,4 @@
 import React from 'react'
-import CucumberQuery from '@cucumber/query'
+import { Query as CucumberQuery } from '@cucumber/query'
 
 export default React.createContext(new CucumberQuery())