+++ react/javascript/src/components/app/QueriesWrapper.tsx
@@ -2,30 +2,25 @@ import React from 'react'
 import GherkinQueryContext from '../../GherkinQueryContext'
 import CucumberQueryContext from '../../CucumberQueryContext'
 import { Query as GherkinQuery } from '@cucumber/gherkin'
-import CucumberQuery from '@cucumber/query'
-import BtoaContext from '../../BtoaContext'
+import { Query as CucumberQuery } from '@cucumber/query'
 
 interface IProps {
   cucumberQuery: CucumberQuery
   gherkinQuery: GherkinQuery
-  btoa: (data: string) => string
 }
 
 const QueriesWrapper: React.FunctionComponent<IProps> = ({
   gherkinQuery,
   cucumberQuery,
-  btoa,
   children,
 }) => {
   return (
     <div className="cucumber-react">
-      <BtoaContext.Provider value={btoa}>
-        <CucumberQueryContext.Provider value={cucumberQuery}>
-          <GherkinQueryContext.Provider value={gherkinQuery}>
-            {children}
-          </GherkinQueryContext.Provider>
-        </CucumberQueryContext.Provider>
-      </BtoaContext.Provider>
+      <CucumberQueryContext.Provider value={cucumberQuery}>
+        <GherkinQueryContext.Provider value={gherkinQuery}>
+          {children}
+        </GherkinQueryContext.Provider>
+      </CucumberQueryContext.Provider>
     </div>
   )
 }