+++ datatable/CHANGELOG.md
@@ -17,6 +17,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [3.3.1] - 2020-03-28
+
+### Fixed
+ * [Java] Improve error messages
+    ([#944](https://github.com/cucumber/cucumber/pull/944)
+      [mpkorstanje])     
+      - `table.asList(String.class)` throw an exception rather then return an empty list   
+
 ## [3.3.0] - 2020-02-06
 
 ### Added
@@ -155,7 +163,8 @@ There are no (1.1.0 and 1.1.1 releases).
 * java: OSGI fixes
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/datatable/v3.3.0...master
+[Unreleased]: https://github.com/cucumber/cucumber/compare/datatable/v3.3.1...master
+[3.3.1]:      https://github.com/cucumber/cucumber/compare/datatable/v3.3.0...datatable/v3.3.1
 [3.3.0]:      https://github.com/cucumber/cucumber/compare/datatable/v3.2.1...datatable/v3.3.0
 [3.2.1]:      https://github.com/cucumber/cucumber/compare/datatable/v3.2.0...datatable/v3.2.1
 [3.2.0]:      https://github.com/cucumber/cucumber/compare/datatable/v3.1.0...datatable/v3.2.0