+++ html-formatter/CHANGELOG.md
@@ -17,6 +17,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+* Fix deprecation waring about `<Wrapper>` (Use `<QueriesWrapper>` instead)
+
+## [4.3.0] - 2020-03-13
+
+### Added
+
+* Ruby implementation
+  ([#931](https://github.com/cucumber/cucumber/pull/931)
+   [vincent-psarga])
+
 ## [4.2.0] - 2020-03-10
 
 ### Added
@@ -25,11 +35,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   ([#922](https://github.com/cucumber/cucumber/pull/922)
    [mpkorstanje])
 * JavaScript: Add a mustache template in the JavaScript npm module that other implementations can use
-* JavaScript: Add `CucumberHtmlStream` (default export), allowing this module to be used as a library (in Cucumber.js) 
+* JavaScript: Add `CucumberHtmlStream` (default export), allowing this module to be used as a library (in Cucumber.js)
 
 ### Changed
 
-* JavaScript: No server side rendering 
+* JavaScript: No server side rendering
   * It's not *really* needed. This does break SEO, but that's not a goal for Cucumber HTML reports.
   * Using both server side rendering and client side rendering results in conflicting versions of the react dom being used during development.
   ([#923](https://github.com/cucumber/cucumber/pull/923)
@@ -146,7 +156,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/formatter/v4.2.0...master
+[Unreleased]: https://github.com/cucumber/cucumber/compare/formatter/v4.3.0...master
+[4.3.0]:      https://github.com/cucumber/cucumber/compare/formatter/v4.2.0...formatter/v4.3.0
 [4.2.0]:      https://github.com/cucumber/cucumber/compare/formatter/v4.1.0...formatter/v4.2.0
 [4.1.0]:      https://github.com/cucumber/cucumber/compare/formatter/v4.0.0...formatter/v4.1.0
 [4.0.0]:      https://github.com/cucumber/cucumber/compare/formatter/v3.2.3...formatter/v4.0.0